+++ libs/filamat/CMakeLists.txt
@@ -9,8 +9,6 @@ set(PUBLIC_HDR_DIR include)
 # ==================================================================================================
 set(HDRS
         include/filamat/Enums.h
-        include/filamat/Includes.h
-        include/filamat/IncludeCallback.h
         include/filamat/MaterialBuilder.h
         include/filamat/Package.h
 )
@@ -40,7 +38,6 @@ set(COMMON_SRCS
         src/shaders/SibGenerator.cpp
         src/shaders/UibGenerator.cpp
         src/Enums.cpp
-        src/Includes.cpp
         src/MaterialBuilder.cpp
         src/MaterialVariants.cpp
 )
@@ -161,8 +158,7 @@ if (IS_HOST_PLATFORM)
             tests/test_filamat.cpp
             tests/test_argBufferFixup.cpp
             tests/test_clipDistanceFixup.cpp
-            tests/test_line_dictionary.cpp
-            tests/test_includes.cpp)
+            tests/test_line_dictionary.cpp)
 
     add_executable(${TARGET} ${SRCS})
     target_include_directories(${TARGET} PRIVATE src)
