+++ libs/filamat/tests/test_filamat.cpp
@@ -20,8 +20,6 @@
 #include "sca/GLSLTools.h"
 #include "shaders/ShaderGenerator.h"
 
-#include "MockIncluder.h"
-
 #include <filamat/Enums.h>
 #include <filamat/MaterialBuilder.h>
 
@@ -53,9 +51,7 @@ std::string shaderWithAllProperties(JobSystem& jobSystem, ShaderStage type,
         filamat::MaterialBuilder::Shading shadingModel = filamat::MaterialBuilder::Shading::LIT,
         filamat::MaterialBuilder::RefractionMode refractionMode = filamat::MaterialBuilder::RefractionMode::NONE,
         filamat::MaterialBuilder::VertexDomain vertexDomain = filamat::MaterialBuilder::VertexDomain::OBJECT) {
-    MockIncluder includer;
-    includer
-            .sourceForInclude("modify_normal.h", "material.normal = vec3(0.8);");
+
     filamat::MaterialBuilder builder;
     builder.material(fragmentCode.c_str());
     builder.materialVertex(vertexCode.c_str());
@@ -68,8 +64,7 @@ std::string shaderWithAllProperties(JobSystem& jobSystem, ShaderStage type,
     MaterialBuilder::PropertyList allProperties;
     std::fill_n(allProperties, MaterialBuilder::MATERIAL_PROPERTIES_COUNT, true);
 
-    // We need to "build" the material to resolve any includes in user code.
-    builder.build(jobSystem);
+    // Note: no need to call builder.build as we are only checking the properties.
 
     return builder.peek(type, {
                     ShaderModel::MOBILE,
@@ -732,25 +727,6 @@ TEST_F(MaterialCompiler, StaticCodeAnalyzerOutputFactor) {
     MaterialBuilder::PropertyList expected{ false };
     expected[size_t(filamat::MaterialBuilder::Property::POST_LIGHTING_COLOR)] = true;
     expected[size_t(filamat::MaterialBuilder::Property::POST_LIGHTING_MIX_FACTOR)] = true;
-    EXPECT_TRUE(PropertyListsMatch(expected, properties));
-}
-
-TEST_F(MaterialCompiler, StaticCodeAnalyzerWithinInclude) {
-    std::string fragmentCode(R"(
-        void material(inout MaterialInputs material) {
-            prepareMaterial(material);
-            #include "modify_normal.h"
-        }
-    )");
-
-    std::string shaderCode = shaderWithAllProperties(*jobSystem, ShaderStage::FRAGMENT,
-            fragmentCode);
-
-    GLSLTools glslTools;
-    MaterialBuilder::PropertyList properties{ false };
-    glslTools.findProperties(ShaderStage::FRAGMENT, shaderCode, properties);
-    MaterialBuilder::PropertyList expected{ false };
-    expected[size_t(filamat::MaterialBuilder::Property::NORMAL)] = true;
     EXPECT_TRUE(PropertyListsMatch(expected, properties));
 }
